fix: Fixed bug related with viewpager and orientation changes.
Fixed a bug where application was crashing on PostDetails and UserProfile screens due to NullPointerException. This exception was being thrown because accessing some elements of fragments of a viewpager after orientation changes  would be null, thus the application would crash with NullPointerException. So, from now on the viewpager fragments are accessed via fragment manager instead.