Add check if current build is release or not.
Add a check in producthunt-wrapper module, if the build is release then don't create build config fields. This is done to prevent an error where the gradle build system checks for debug variant even when release specific task is fired. The error in this scenario was that while running assembleRelease task for producthunt-wrapper, it will check for external properties inside debug variant also, which would result in a crash as system won't be able to find these properties in CI environment as config.gradle is not added to git.